Move util schemas new file
