+++ src/applications/disability-benefits/all-claims/pages/secondaryIncidentAuthorities.js
@@ -2,7 +2,7 @@ import fullSchema from 'vets-json-schema/dist/21-526EZ-ALLCLAIMS-schema.json';
 import AuthorityField from '../components/AuthorityField';
 import { ptsd781aNameTitle } from '../content/ptsdClassification';
 import { PtsdAssaultAuthoritiesDescription } from '../content/ptsdAssaultAuthorities';
-import { addressUISchema } from '../utils';
+import { addressUISchema } from '../utils/schemas';
 
 const {
   sources,
