+++ src/applications/disability-benefits/all-claims/pages/secondaryUploadSources.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
+import fullSchema from 'vets-json-schema/dist/21-526EZ-ALLCLAIMS-schema.json';
 
-import { ancillaryFormUploadUi } from '../utils';
-
+import { ancillaryFormUploadUi } from '../utils/schemas';
 import { UploadDescription } from '../content/fileUploadDescriptions';
 import { ptsd781aNameTitle } from '../content/ptsdClassification';
-import fullSchema from 'vets-json-schema/dist/21-526EZ-ALLCLAIMS-schema.json';
 
 const { secondaryAttachment } = fullSchema.properties;
