+++ src/applications/disability-benefits/all-claims/pages/uploadPersonalPtsdDocuments.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { UploadDescription } from '../content/fileUploadDescriptions';
 import { ptsd781aNameTitle } from '../content/ptsdClassification';
-import { ancillaryFormUploadUi, getAttachmentsSchema } from '../utils';
+import { ancillaryFormUploadUi, getAttachmentsSchema } from '../utils/schemas';
 
 const PTSD_781A_ATTACHMENT_ID = 'L229';
