+++ src/applications/disability-benefits/all-claims/tests/utils/schemas.unit.spec.js
@@ -1,0 +1,104 @@
+import { expect } from 'chai';
+
+import {
+  makeSchemaForNewDisabilities,
+  makeSchemaForRatedDisabilities,
+  makeSchemaForAllDisabilities,
+} from '../../utils/schemas';
+
+describe('makeSchemaForNewDisabilities', () => {
+  it('should return schema with downcased keynames', () => {
+    const formData = {
+      newDisabilities: [
+        {
+          condition: 'Ptsd personal trauma',
+        },
+      ],
+    };
+    expect(makeSchemaForNewDisabilities(formData)).to.eql({
+      properties: {
+        ptsdpersonaltrauma: {
+          title: 'Ptsd Personal Trauma',
+          type: 'boolean',
+        },
+      },
+    });
+  });
+
+  it('should return correct schema when periods used', () => {
+    const formData = {
+      newDisabilities: [
+        {
+          condition: 'period. Period.',
+        },
+      ],
+    };
+    expect(makeSchemaForNewDisabilities(formData)).to.eql({
+      properties: {
+        periodperiod: {
+          title: 'Period. Period.',
+          type: 'boolean',
+        },
+      },
+    });
+  });
+});
+
+describe('makeSchemaForRatedDisabilities', () => {
+  it('should return schema for selected disabilities only', () => {
+    const formData = {
+      ratedDisabilities: [
+        {
+          name: 'Ptsd personal trauma',
+          'view:selected': false,
+        },
+        {
+          name: 'Diabetes mellitus',
+          'view:selected': true,
+        },
+      ],
+    };
+    expect(makeSchemaForRatedDisabilities(formData)).to.eql({
+      properties: {
+        diabetesmellitus: {
+          title: 'Diabetes Mellitus',
+          type: 'boolean',
+        },
+      },
+    });
+  });
+});
+
+describe('makeSchemaForAllDisabilities', () => {
+  it('should return schema for all (selected) disabilities', () => {
+    const formData = {
+      ratedDisabilities: [
+        {
+          name: 'Ptsd personal trauma',
+          'view:selected': false,
+        },
+        {
+          name: 'Diabetes mellitus',
+          'view:selected': true,
+        },
+      ],
+      newDisabilities: [
+        {
+          condition: 'A new Condition.',
+        },
+      ],
+    };
+    expect(makeSchemaForAllDisabilities(formData)).to.eql({
+      properties: {
+        diabetesmellitus: {
+          title: 'Diabetes Mellitus',
+          type: 'boolean',
+        },
+        anewcondition: {
+          title: 'A New Condition.',
+          type: 'boolean',
+        },
+      },
+    });
+  });
+});
