+++ src/applications/disability-benefits/all-claims/tests/utils/utils.unit.spec.jsx
@@ -10,9 +10,6 @@ import {
   CHAR_LIMITS,
 } from '../../constants';
 import {
-  makeSchemaForNewDisabilities,
-  makeSchemaForRatedDisabilities,
-  makeSchemaForAllDisabilities,
   capitalizeEachWord,
   fieldsHaveInput,
   hasGuardOrReservePeriod,
@@ -180,103 +177,6 @@ describe('526 helpers', () => {
     });
     it('should return null when name is not a string', () => {
       expect(capitalizeEachWord(249481)).to.equal(null);
-    });
-  });
-
-  describe('makeSchemaForNewDisabilities', () => {
-    it('should return schema with downcased keynames', () => {
-      const formData = {
-        newDisabilities: [
-          {
-            condition: 'Ptsd personal trauma',
-          },
-        ],
-      };
-      expect(makeSchemaForNewDisabilities(formData)).to.eql({
-        properties: {
-          ptsdpersonaltrauma: {
-            title: 'Ptsd Personal Trauma',
-            type: 'boolean',
-          },
-        },
-      });
-    });
-
-    it('should return correct schema when periods used', () => {
-      const formData = {
-        newDisabilities: [
-          {
-            condition: 'period. Period.',
-          },
-        ],
-      };
-      expect(makeSchemaForNewDisabilities(formData)).to.eql({
-        properties: {
-          periodperiod: {
-            title: 'Period. Period.',
-            type: 'boolean',
-          },
-        },
-      });
-    });
-  });
-
-  describe('makeSchemaForRatedDisabilities', () => {
-    it('should return schema for selected disabilities only', () => {
-      const formData = {
-        ratedDisabilities: [
-          {
-            name: 'Ptsd personal trauma',
-            'view:selected': false,
-          },
-          {
-            name: 'Diabetes mellitus',
-            'view:selected': true,
-          },
-        ],
-      };
-      expect(makeSchemaForRatedDisabilities(formData)).to.eql({
-        properties: {
-          diabetesmellitus: {
-            title: 'Diabetes Mellitus',
-            type: 'boolean',
-          },
-        },
-      });
-    });
-  });
-
-  describe('makeSchemaForAllDisabilities', () => {
-    it('should return schema for all (selected) disabilities', () => {
-      const formData = {
-        ratedDisabilities: [
-          {
-            name: 'Ptsd personal trauma',
-            'view:selected': false,
-          },
-          {
-            name: 'Diabetes mellitus',
-            'view:selected': true,
-          },
-        ],
-        newDisabilities: [
-          {
-            condition: 'A new Condition.',
-          },
-        ],
-      };
-      expect(makeSchemaForAllDisabilities(formData)).to.eql({
-        properties: {
-          diabetesmellitus: {
-            title: 'Diabetes Mellitus',
-            type: 'boolean',
-          },
-          anewcondition: {
-            title: 'A New Condition.',
-            type: 'boolean',
-          },
-        },
-      });
     });
   });
